Implemented a basic version of Queue
Queue is a Stack-esque array that holds tasks before execution and
executes them in parallel. It is designed to efficiently perform
optimizations over execution of tasks based on a bunch of factors but
for now features a simple execution stack with the ability to batch a
bunch of executions using setImmediate.